Working preliminary profile needs to have information added after foundation for information is created. Icon added a long with temporary image.

+++ BACHome/CreateAccountViewController.m
@@ -107,7 +107,7 @@
                 } else {
                     NSString *errorString = [[error userInfo] objectForKey:@"error"];
                     NSInteger code = [error code];
-                    NSLog(@"Error: %@, %ld", errorString, code);
+                    NSLog(@"Error: %@, %ld", errorString, (long)code);
                     if (code == 125) {
                         [_activityWheel stopAnimating];
                         [_emailField becomeFirstResponder];
